voice_api/requirements.txt
# Voice API Requirements

# Core dependencies
google-genai>=0.2.0
pyaudio>=0.2.13

# Voice Activity Detection (WebRTC-based, production-grade)
webrtcvad>=2.0.10


